Move multiple email and mobile feature status check to user management utils
